@@ -8,13 +8,14 @@ import (
 	"reflect"
 
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
-	"github.com/pulumi/pulumi/sdk/v3/go/pulumix"
 	"github.com/pulumiverse/pulumi-time/sdk/go/time/internal"
 )
 
 // ## Example Usage
+//
 // ### Basic Usage
 //
+// <!--Start PulumiCodeChooser -->
 // ```go
 // package main
 //
@@ -38,26 +39,23 @@ import (
 //	}
 //
 // ```
+// <!--End PulumiCodeChooser -->
 //
 // ## Import
 //
 // This resource can be imported using the base UTC RFC3339 value and rotation years, months, days, hours, and minutes, separated by commas (`,`), e.g. for 30 days
 //
 // ```sh
-//
-//	$ pulumi import time:index/rotating:Rotating example 2020-02-12T06:36:13Z,0,0,30,0,0
-//
+// $ pulumi import time:index/rotating:Rotating example 2020-02-12T06:36:13Z,0,0,30,0,0
 // ```
 //
-//	Otherwise, to import with the rotation RFC3339 value, the base UTC RFC3339 value and rotation UTC RFC3339 value, separated by commas (`,`), e.g.
+// Otherwise, to import with the rotation RFC3339 value, the base UTC RFC3339 value and rotation UTC RFC3339 value, separated by commas (`,`), e.g.
 //
 // ```sh
-//
-//	$ pulumi import time:index/rotating:Rotating example 2020-02-12T06:36:13Z,2020-02-13T06:36:13Z
-//
+// $ pulumi import time:index/rotating:Rotating example 2020-02-12T06:36:13Z,2020-02-13T06:36:13Z
 // ```
 //
-//	The `triggers` argument cannot be imported.
+// The `triggers` argument cannot be imported.
 type Rotating struct {
 	pulumi.CustomResourceState
 
@@ -254,12 +252,6 @@ func (i *Rotating) ToRotatingOutputWithContext(ctx context.Context) RotatingOutp
 	return pulumi.ToOutputWithContext(ctx, i).(RotatingOutput)
 }
 
-func (i *Rotating) ToOutput(ctx context.Context) pulumix.Output[*Rotating] {
-	return pulumix.Output[*Rotating]{
-		OutputState: i.ToRotatingOutputWithContext(ctx).OutputState,
-	}
-}
-
 // RotatingArrayInput is an input type that accepts RotatingArray and RotatingArrayOutput values.
 // You can construct a concrete instance of `RotatingArrayInput` via:
 //
@@ -283,12 +275,6 @@ func (i RotatingArray) ToRotatingArrayOutput() RotatingArrayOutput {
 
 func (i RotatingArray) ToRotatingArrayOutputWithContext(ctx context.Context) RotatingArrayOutput {
 	return pulumi.ToOutputWithContext(ctx, i).(RotatingArrayOutput)
-}
-
-func (i RotatingArray) ToOutput(ctx context.Context) pulumix.Output[[]*Rotating] {
-	return pulumix.Output[[]*Rotating]{
-		OutputState: i.ToRotatingArrayOutputWithContext(ctx).OutputState,
-	}
 }
 
 // RotatingMapInput is an input type that accepts RotatingMap and RotatingMapOutput values.
@@ -316,12 +302,6 @@ func (i RotatingMap) ToRotatingMapOutputWithContext(ctx context.Context) Rotatin
 	return pulumi.ToOutputWithContext(ctx, i).(RotatingMapOutput)
 }
 
-func (i RotatingMap) ToOutput(ctx context.Context) pulumix.Output[map[string]*Rotating] {
-	return pulumix.Output[map[string]*Rotating]{
-		OutputState: i.ToRotatingMapOutputWithContext(ctx).OutputState,
-	}
-}
-
 type RotatingOutput struct{ *pulumi.OutputState }
 
 func (RotatingOutput) ElementType() reflect.Type {
@@ -334,12 +314,6 @@ func (o RotatingOutput) ToRotatingOutput() RotatingOutput {
 
 func (o RotatingOutput) ToRotatingOutputWithContext(ctx context.Context) RotatingOutput {
 	return o
-}
-
-func (o RotatingOutput) ToOutput(ctx context.Context) pulumix.Output[*Rotating] {
-	return pulumix.Output[*Rotating]{
-		OutputState: o.OutputState,
-	}
 }
 
 // Number day of timestamp.
@@ -431,12 +405,6 @@ func (o RotatingArrayOutput) ToRotatingArrayOutputWithContext(ctx context.Contex
 	return o
 }
 
-func (o RotatingArrayOutput) ToOutput(ctx context.Context) pulumix.Output[[]*Rotating] {
-	return pulumix.Output[[]*Rotating]{
-		OutputState: o.OutputState,
-	}
-}
-
 func (o RotatingArrayOutput) Index(i pulumi.IntInput) RotatingOutput {
 	return pulumi.All(o, i).ApplyT(func(vs []interface{}) *Rotating {
 		return vs[0].([]*Rotating)[vs[1].(int)]
@@ -455,12 +423,6 @@ func (o RotatingMapOutput) ToRotatingMapOutput() RotatingMapOutput {
 
 func (o RotatingMapOutput) ToRotatingMapOutputWithContext(ctx context.Context) RotatingMapOutput {
 	return o
-}
-
-func (o RotatingMapOutput) ToOutput(ctx context.Context) pulumix.Output[map[string]*Rotating] {
-	return pulumix.Output[map[string]*Rotating]{
-		OutputState: o.OutputState,
-	}
 }
 
 func (o RotatingMapOutput) MapIndex(k pulumi.StringInput) RotatingOutput {
