@@ -8,13 +8,14 @@ import (
 	"reflect"
 
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
-	"github.com/pulumi/pulumi/sdk/v3/go/pulumix"
 	"github.com/pulumiverse/pulumi-time/sdk/go/time/internal"
 )
 
 // ## Example Usage
+//
 // ### Basic Usage
 //
+// <!--Start PulumiCodeChooser -->
 // ```go
 // package main
 //
@@ -37,8 +38,11 @@ import (
 //	}
 //
 // ```
+// <!--End PulumiCodeChooser -->
+//
 // ### Triggers Usage
 //
+// <!--Start PulumiCodeChooser -->
 // ```go
 // package main
 //
@@ -76,18 +80,17 @@ import (
 //	}
 //
 // ```
+// <!--End PulumiCodeChooser -->
 //
 // ## Import
 //
 // This resource can be imported using the UTC RFC3339 value, e.g.
 //
 // ```sh
-//
-//	$ pulumi import time:index/static:Static example 2020-02-12T06:36:13Z
-//
+// $ pulumi import time:index/static:Static example 2020-02-12T06:36:13Z
 // ```
 //
-//	The `triggers` argument cannot be imported.
+// The `triggers` argument cannot be imported.
 type Static struct {
 	pulumi.CustomResourceState
 
@@ -224,12 +227,6 @@ func (i *Static) ToStaticOutputWithContext(ctx context.Context) StaticOutput {
 	return pulumi.ToOutputWithContext(ctx, i).(StaticOutput)
 }
 
-func (i *Static) ToOutput(ctx context.Context) pulumix.Output[*Static] {
-	return pulumix.Output[*Static]{
-		OutputState: i.ToStaticOutputWithContext(ctx).OutputState,
-	}
-}
-
 // StaticArrayInput is an input type that accepts StaticArray and StaticArrayOutput values.
 // You can construct a concrete instance of `StaticArrayInput` via:
 //
@@ -253,12 +250,6 @@ func (i StaticArray) ToStaticArrayOutput() StaticArrayOutput {
 
 func (i StaticArray) ToStaticArrayOutputWithContext(ctx context.Context) StaticArrayOutput {
 	return pulumi.ToOutputWithContext(ctx, i).(StaticArrayOutput)
-}
-
-func (i StaticArray) ToOutput(ctx context.Context) pulumix.Output[[]*Static] {
-	return pulumix.Output[[]*Static]{
-		OutputState: i.ToStaticArrayOutputWithContext(ctx).OutputState,
-	}
 }
 
 // StaticMapInput is an input type that accepts StaticMap and StaticMapOutput values.
@@ -286,12 +277,6 @@ func (i StaticMap) ToStaticMapOutputWithContext(ctx context.Context) StaticMapOu
 	return pulumi.ToOutputWithContext(ctx, i).(StaticMapOutput)
 }
 
-func (i StaticMap) ToOutput(ctx context.Context) pulumix.Output[map[string]*Static] {
-	return pulumix.Output[map[string]*Static]{
-		OutputState: i.ToStaticMapOutputWithContext(ctx).OutputState,
-	}
-}
-
 type StaticOutput struct{ *pulumi.OutputState }
 
 func (StaticOutput) ElementType() reflect.Type {
@@ -304,12 +289,6 @@ func (o StaticOutput) ToStaticOutput() StaticOutput {
 
 func (o StaticOutput) ToStaticOutputWithContext(ctx context.Context) StaticOutput {
 	return o
-}
-
-func (o StaticOutput) ToOutput(ctx context.Context) pulumix.Output[*Static] {
-	return pulumix.Output[*Static]{
-		OutputState: o.OutputState,
-	}
 }
 
 // Number day of timestamp.
@@ -371,12 +350,6 @@ func (o StaticArrayOutput) ToStaticArrayOutputWithContext(ctx context.Context) S
 	return o
 }
 
-func (o StaticArrayOutput) ToOutput(ctx context.Context) pulumix.Output[[]*Static] {
-	return pulumix.Output[[]*Static]{
-		OutputState: o.OutputState,
-	}
-}
-
 func (o StaticArrayOutput) Index(i pulumi.IntInput) StaticOutput {
 	return pulumi.All(o, i).ApplyT(func(vs []interface{}) *Static {
 		return vs[0].([]*Static)[vs[1].(int)]
@@ -395,12 +368,6 @@ func (o StaticMapOutput) ToStaticMapOutput() StaticMapOutput {
 
 func (o StaticMapOutput) ToStaticMapOutputWithContext(ctx context.Context) StaticMapOutput {
 	return o
-}
-
-func (o StaticMapOutput) ToOutput(ctx context.Context) pulumix.Output[map[string]*Static] {
-	return pulumix.Output[map[string]*Static]{
-		OutputState: o.OutputState,
-	}
 }
 
 func (o StaticMapOutput) MapIndex(k pulumi.StringInput) StaticOutput {
