@@ -213,8 +213,10 @@ class Static(pulumi.CustomResource):
                  __props__=None):
         """
         ## Example Usage
+
         ### Basic Usage
 
+        <!--Start PulumiCodeChooser -->
         ```python
         import pulumi
         import pulumiverse_time as time
@@ -222,8 +224,11 @@ class Static(pulumi.CustomResource):
         example = time.Static("example")
         pulumi.export("currentTime", example.rfc3339)
         ```
+        <!--End PulumiCodeChooser -->
+
         ### Triggers Usage
 
+        <!--Start PulumiCodeChooser -->
         ```python
         import pulumi
         import pulumi_aws as aws
@@ -239,16 +244,17 @@ class Static(pulumi.CustomResource):
             })
         # ... (other aws_instance arguments) ...
         ```
+        <!--End PulumiCodeChooser -->
 
         ## Import
 
         This resource can be imported using the UTC RFC3339 value, e.g.
 
         ```sh
-         $ pulumi import time:index/static:Static example 2020-02-12T06:36:13Z
+        $ pulumi import time:index/static:Static example 2020-02-12T06:36:13Z
         ```
 
-         The `triggers` argument cannot be imported.
+        The `triggers` argument cannot be imported.
 
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
@@ -263,8 +269,10 @@ class Static(pulumi.CustomResource):
                  opts: Optional[pulumi.ResourceOptions] = None):
         """
         ## Example Usage
+
         ### Basic Usage
 
+        <!--Start PulumiCodeChooser -->
         ```python
         import pulumi
         import pulumiverse_time as time
@@ -272,8 +280,11 @@ class Static(pulumi.CustomResource):
         example = time.Static("example")
         pulumi.export("currentTime", example.rfc3339)
         ```
+        <!--End PulumiCodeChooser -->
+
         ### Triggers Usage
 
+        <!--Start PulumiCodeChooser -->
         ```python
         import pulumi
         import pulumi_aws as aws
@@ -289,16 +300,17 @@ class Static(pulumi.CustomResource):
             })
         # ... (other aws_instance arguments) ...
         ```
+        <!--End PulumiCodeChooser -->
 
         ## Import
 
         This resource can be imported using the UTC RFC3339 value, e.g.
 
         ```sh
-         $ pulumi import time:index/static:Static example 2020-02-12T06:36:13Z
+        $ pulumi import time:index/static:Static example 2020-02-12T06:36:13Z
         ```
 
-         The `triggers` argument cannot be imported.
+        The `triggers` argument cannot be imported.
 
         :param str resource_name: The name of the resource.
         :param StaticArgs args: The arguments to use to populate this resource's properties.
